+++ test/core_tests/task_registry_test.py
@@ -1,5 +1,6 @@
 """Unit test for task registry."""
 import asyncio
+import sys
 
 from xknx import XKNX
 from xknx.core import XknxConnectionState
@@ -124,3 +125,26 @@ class TestTaskRegistry:
         await asyncio.sleep(0)  # iterate loop to cancel task
         assert self.test == 0
         assert len(xknx.connection_manager._connection_state_changed_cbs) == 0
+
+    async def test_background(self, time_travel):
+        """Test running background task."""
+        TIME = 10
+
+        async def callback() -> None:
+            """Do nothing."""
+            await asyncio.sleep(TIME)
+
+        xknx = XKNX()
+        xknx.task_registry.background(callback())
+        assert len(xknx.task_registry._background_task) == 1
+        task = next(iter(xknx.task_registry._background_task))
+        refs = sys.getrefcount(task)
+        assert refs == 4
+        assert not task.done()
+
+        # after task is finished it should remove itself from the background registry
+        await time_travel(TIME)
+        assert len(xknx.task_registry._background_task) == 0
+        assert task.done()
+        refs = sys.getrefcount(task)
+        assert refs == 2